refactor(v2): split feature contracts by product domain
