+++ src/v2/shared/features/prompts.ts
@@ -1,0 +1,9 @@
+export interface ManagedPrompt {
+  id: string;
+  name: string;
+  content: string;
+  description?: string;
+  enabled: boolean;
+  createdAt?: number;
+  updatedAt?: number;
+}
