+++ src/v2/shared/features/settings.ts
@@ -1,0 +1,4 @@
+export type FeatureSettings = Record<string, unknown> & {
+  skillSyncMethod?: "auto" | "symlink" | "copy";
+  skillStorageLocation?: "fyagent" | "unified";
+};
